Restructure and added Bootstrap Bootstrap is added using bower, and is automatically downloaded upon using npm install.
