+++ frontend/app.js
@@ -1,5 +1,5 @@
-var server = require('./router');
-var logger = require('./logger');
+var server = require('./modules/router');
+var logger = require('./modules/logger');
 
 server.listen(3000, function () {
 	// Once the server starts, reset the logfile.
