--- frontend/app/web/index.html
@@ -1,12 +1,0 @@
-<!DOCTYPE html>
-<html>
-<head>
-<meta charset="UTF-8">
-<title>Preset Control Web Application</title>
-</head>
-
-<body>
-Hello World!
-</body>
-
-</html>
