+++ frontend/public/css/style.css
@@ -1,0 +1,5 @@
+/* navbar */
+.navbar-default {
+    background-color: #666666;
+    border-color: #E7E7E7;
+}
