+++ frontend/public/index.html
@@ -1,0 +1,26 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta http-equiv="X-UA-Compatible" content="IE=edge">
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+	
+	<title>Preset Control Web Application</title>
+
+	<!-- Bootstrap -->
+    <link href="bower_components/bootstrap/dist/css/bootstrap.min.css" rel="stylesheet">
+    <!-- Bootstrap CSS overwrites -->
+    <link href="public/css/style.css" rel="stylesheet">
+</head>
+
+<body>
+<script src="bower_components/jquery/dist/jquery.min.js"></script>
+<script src="bower_components/bootstrap/dist/js/bootstrap.min.js"></script>
+<nav class="navbar navbar-default navbar-fixed-top">
+  <div class="container">
+     <a class="navbar-brand" href="#">PolyCast WebInterface</a>
+  </div>
+</nav>
+</body>
+
+</html>
